requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
streamlit==1.42.2
requests==2.32.3
python-dotenv==1.0.1
groq==0.19.0
pydantic==2.10.6
